+++ nodeup/pkg/model/BUILD.bazel
@@ -12,7 +12,6 @@ go_library(
         "docker.go",
         "etcd.go",
         "etcd_manager_tls.go",
-        "etcd_tls.go",
         "file_assets.go",
         "firewall.go",
         "hooks.go",
