+++ nodeup/pkg/model/networking/BUILD.bazel
@@ -3,6 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = [
+        "calico.go",
         "cilium.go",
         "flannel.go",
         "kube_router.go",
